requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.20
transformers
torch
PyMuPDF==1.25.1
python-docx==1.1.2
pytesseract==0.3.13
Pillow==11.0.0
python-dotenv==1.0.1
pydantic==2.10.3
